// src/interfaces/conversation.ts
export interface Conversation {
    id?: number;
    number: string;
    role: string;
    content: string;
    code: string;
    // date: Date;
    // start_date: Date
    // end_date: Date;
    // archived?: boolean;
    // conversation_id?: number;
}

// export interface Conversation {
//     id?: number;
//     number: number;
//     start_date: Date
//     end_date: Date;
//     archived?: boolean;
// }
// export interface Chat {
//     id?: number;
//     conversation_id?: number;
//     role: string;
//     content: string;
//     date: Date
//     code: string;
// }